doc: Improve naming and documentation for code around IConfiguration.
